Move file to mapbox dir
