+++ src/components/mapbox/mapstyles.ts
@@ -1,0 +1,11 @@
+export const MAPBOX_STYLE = {
+  STANDARD: "mapbox://styles/mapbox/standard",
+  STREETS: "mapbox://styles/mapbox/streets-v12",
+  OUTDOORS: "mapbox://styles/mapbox/outdoors-v12",
+  LIGHT: "mapbox://styles/mapbox/light-v11",
+  DARK: "mapbox://styles/mapbox/dark-v11",
+  SAT: "mapbox://styles/mapbox/satellite-v9",
+  SAT_STREETS: "mapbox://styles/mapbox/satellite-streets-v12",
+  NAV_DAY: "mapbox://styles/mapbox/navigation-day-v1",
+  NAV_NIGHT: "mapbox://styles/mapbox/navigation-night-v1",
+};
